Split sequence generators into individual files
