+++ lib/iterate.js
@@ -6,13 +6,13 @@ const isIterable = require("./isiterable");
 function defineSequence(name) {
 	return function(...args) {
 		return proxy.makeSequence(name, ...args);
-	}
+	};
 }
 
 function defineChainableIterator(name) {
 	return function(...args) {
 		return proxy.makeChainableIterator(name, ...args);
-	}
+	};
 }
 
 let genUtil = {
@@ -25,7 +25,7 @@ for (sequence of Reflect.ownKeys(sequences)) {
 		enumerable: true,
 		writable: true,
 		value: defineSequence(sequence)
-	})
+	});
 }
 
 for (chainable of Reflect.ownKeys(chainables)) {
@@ -34,7 +34,7 @@ for (chainable of Reflect.ownKeys(chainables)) {
 		enumerable: true,
 		writable: true,
 		value: defineChainableIterator(chainable)
-	})
+	});
 }
 
 module.exports = genUtil;
