+++ lib/sequences/collatz.js
@@ -1,0 +1,13 @@
+const { series } = require("./series");
+const { takeUntil } = require("../chainables");
+
+function* collatz(n) {
+	if (n < 1 || n % 1) return;
+
+	let f = n => (n % 2 ? n * 3 + 1 : n / 2);
+	let s = series(f, n);
+	let predicate = n => n === 1;
+	yield* takeUntil(s, predicate);
+}
+
+module.exports = collatz;
