+++ lib/sequences/fizzBuzz.js
@@ -1,0 +1,7 @@
+const { fuzz } = require("../chainables");
+
+function* fizzBuzz() {
+	yield* fuzz(counter(1));
+}
+
+module.exports = fizzBuzz;
