+++ lib/sequences/series.js
@@ -1,0 +1,23 @@
+/**
+ *
+ * @param f
+ * @param n
+ */
+function* series(f, ...n) {
+	if (n.length === 0) {
+		n = [1];
+	}
+
+	for (let value of n) {
+		yield value;
+	}
+
+	while (true) {
+		let next = f(...n);
+		n.shift();
+		n.push(next);
+		yield next;
+	}
+}
+
+module.exports = series;
