+++ lib/sequences/symbols.js
@@ -1,0 +1,7 @@
+function* symbols(name) {
+	while (true) {
+		yield Symbol(name);
+	}
+}
+
+module.exports = symbols;
